Split ProfileFactoriesConfiguration into per-feature configurations
- Replace the 718-line, 44-bean ProfileFactoriesConfiguration with four
  feature-scoped classes along the existing holder-record seams:
  ProfileVisualizationConfiguration (6 beans), ProfileAnalysisConfiguration
  (8), ProfileJvmInsightConfiguration (21), ProfileCoreConfiguration (9)
- Move the RECORDINGS_PATH/PROFILES_PATH qualifier constants to the
  ProfilesConfiguration aggregator and update core-microscope references
- Adopt the shared Exceptions factory in profile-management: convert the
  seven scattered raw RuntimeException throws (auto-analysis, recording
  info parser, heap-dump upload/session/report, perf counters, profile
  data initializer) to Exceptions.internal with preserved causes
