+++ jeffrey-microscope/profiles/profile-management/src/main/java/cafe/jeffrey/profile/configuration/ProfilesConfiguration.java
@@ -1,3 +1,21 @@
+/*
+ * Jeffrey
+ * Copyright (C) 2026 Petr Bouda
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 package cafe.jeffrey.profile.configuration;
 
 import org.springframework.context.annotation.Import;
@@ -15,8 +33,14 @@ import cafe.jeffrey.profile.heapdump.oql.config.OqlEngineConfiguration;
         DuckDbMcpConfiguration.class,
         HeapDumpMcpConfiguration.class,
         OqlEngineConfiguration.class,
-        ProfileFactoriesConfiguration.class,
+        ProfileCoreConfiguration.class,
+        ProfileVisualizationConfiguration.class,
+        ProfileAnalysisConfiguration.class,
+        ProfileJvmInsightConfiguration.class,
         ProfileCustomFactoriesConfiguration.class
 })
 public class ProfilesConfiguration {
+
+    public static final String RECORDINGS_PATH = "recordings-path";
+    public static final String PROFILES_PATH = "profiles-path";
 }
