+++ jeffrey-microscope/profiles/profile-management/src/main/java/cafe/jeffrey/profile/manager/heapdump/HeapDumpSessionTemplate.java
@@ -21,6 +21,7 @@ package cafe.jeffrey.profile.manager.heapdump;
 import cafe.jeffrey.profile.heapdump.parser.HeapDumpIndexPaths;
 import cafe.jeffrey.profile.heapdump.parser.HeapDumpSession;
 import cafe.jeffrey.profile.manager.AdditionalFilesManager;
+import cafe.jeffrey.shared.common.exception.Exceptions;
 import cafe.jeffrey.shared.common.model.ProfileInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -73,7 +74,7 @@ public final class HeapDumpSessionTemplate {
         } catch (IOException | SQLException e) {
             LOG.warn("Heap dump operation failed: profileId={} path={} error={}",
                     profileInfo.id(), heapPath.get(), e.getMessage());
-            throw new RuntimeException("Heap dump operation failed: " + e.getMessage(), e);
+            throw Exceptions.internal("Heap dump operation failed: " + e.getMessage(), e);
         }
     }
